sayma2: JESD204 synchronization
