+++ artiq/firmware/libproto_artiq/drtioaux_proto.rs
@@ -54,8 +54,8 @@ pub enum Packet {
     SpiReadReply { succeeded: bool, data: u32 },
     SpiBasicReply { succeeded: bool },
 
-    JdacBasicRequest { destination: u8, dacno: u8, reqno: u8 },
-    JdacBasicReply { succeeded: bool },
+    JdacBasicRequest { destination: u8, dacno: u8, reqno: u8, param: u8 },
+    JdacBasicReply { succeeded: bool, retval: u8 },
 }
 
 impl Packet {
@@ -185,9 +185,11 @@ impl Packet {
                 destination: reader.read_u8()?,
                 dacno: reader.read_u8()?,
                 reqno: reader.read_u8()?,
+                param: reader.read_u8()?,
             },
             0xa1 => Packet::JdacBasicReply {
-                succeeded: reader.read_bool()?
+                succeeded: reader.read_bool()?,
+                retval: reader.read_u8()?
             },
 
             ty => return Err(Error::UnknownPacket(ty))
@@ -342,15 +344,17 @@ impl Packet {
                 writer.write_bool(succeeded)?;
             },
 
-            Packet::JdacBasicRequest { destination, dacno, reqno } => {
+            Packet::JdacBasicRequest { destination, dacno, reqno, param } => {
                 writer.write_u8(0xa0)?;
                 writer.write_u8(destination)?;
                 writer.write_u8(dacno)?;
                 writer.write_u8(reqno)?;
+                writer.write_u8(param)?;
             }
-            Packet::JdacBasicReply { succeeded } => {
+            Packet::JdacBasicReply { succeeded, retval } => {
                 writer.write_u8(0xa1)?;
                 writer.write_bool(succeeded)?;
+                writer.write_u8(retval)?;
             },
         }
         Ok(())
