+++ artiq/firmware/satman/jdac_requests.rs
@@ -2,3 +2,7 @@ pub const INIT: u8              = 0x00;
 pub const PRINT_STATUS: u8      = 0x01;
 pub const PRBS: u8              = 0x02;
 pub const STPL: u8              = 0x03;
+
+pub const SYSREF_DELAY_DAC: u8  = 0x10;
+pub const SYSREF_SLIP: u8       = 0x11;
+pub const SYNC: u8              = 0x12;
